add solution file
