--- ModConfig.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using Microsoft.Xna.Framework;
-using StardewModdingAPI;
-using StardewModdingAPI.Events;
-using StardewModdingAPI.Utilities;
-using StardewValley;
-
-namespace Mopsy_Ranch_Livin
-{
-    internal class ModConfig
-    {
-        /// <summary>The farm name suffix to use.</summary>
-        public string Suffix { get; set; }
-    }
-}
