+++ Mopsy-Ranch-Livin'.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28407.52
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mopsy-Ranch-Livin'", "Mopsy-Ranch-Livin'\Mopsy-Ranch-Livin'.csproj", "{3EF556D6-575B-452B-BB69-B012CE133FE6}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3EF556D6-575B-452B-BB69-B012CE133FE6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3EF556D6-575B-452B-BB69-B012CE133FE6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3EF556D6-575B-452B-BB69-B012CE133FE6}.Debug|x86.ActiveCfg = Debug|x86
+		{3EF556D6-575B-452B-BB69-B012CE133FE6}.Debug|x86.Build.0 = Debug|x86
+		{3EF556D6-575B-452B-BB69-B012CE133FE6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3EF556D6-575B-452B-BB69-B012CE133FE6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3EF556D6-575B-452B-BB69-B012CE133FE6}.Release|x86.ActiveCfg = Release|x86
+		{3EF556D6-575B-452B-BB69-B012CE133FE6}.Release|x86.Build.0 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A158BFF5-D230-40D0-A2EF-1CE54ADB382E}
+	EndGlobalSection
+EndGlobal
